requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
openai==1.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
